addLocation

Function translates (if possible) cell numbers into coordinates or coordinates into cell numbers and adds them to the spatial dimension of the object.

Usage

addLocation(x, fillMissing = NULL, naCellNumber = 0, format = "both")

Arguments

x

magpie object to be enriched with spatial information. Currently only works for 0.5 degree data sets with 59199 or 67420 cells or a subset of it.

fillMissing

if NULL cells missing from the total 59199 are just being ignore. If set to a value missing cells will be added with this value (e.g. all set to 0 if fillMissing is 0)

naCellNumber

number or string given as cell number for cells which do not have an assigned cell number in the 59199 mapping

format

either "both" or "iso.grid". When set to "iso.grid", the function will remove the pre-existing spatial dimensions after adding the coordinates' sorted iso.grid location. This argument is ignored for input objects initially containing iso.grid and LPJ datasets.

Value

magpie object with added spatial subdimensions

Author

Jan Philipp Dietrich, Felicitas Beier, Michael Crawford

Examples

 map <- Cell2Country()
 x <- y <- z <- population_magpie

 getCoords(x) <- map[100:109, c("lon", "lat")]
 getItems(x, "i") <- NULL
 addLocation(x)
#> , , scenario = A2
#> 
#>                       t
#> x.y.country.cell           y1995   y2005   y2015   y2025   y2035
#>   -14p75.20p25.MRT.100  552.6664  696.44  889.18 1124.11 1389.33
#>   -14p75.20p75.MRT.101 1280.6350 1429.53 1518.46 1592.09 1640.95
#>   -14p75.21p25.MRT.102  554.4384  582.36  593.76  605.27  614.58
#>   -14p75.21p75.ESH.103  276.3431  295.38  302.62  308.59  313.30
#>   -14p75.22p25.ESH.104  451.9981  558.29  646.02  733.13  812.69
#>   -14p75.22p75.ESH.105  277.7437  390.18  489.22  596.13  698.33
#>   -14p75.23p25.ESH.106  292.1132  326.09  353.25  382.53  409.44
#>   -14p75.23p75.ESH.107  133.7772  152.00  155.27  157.35  158.81
#>   -14p75.24p25.ESH.108  383.2277  534.73  604.94  668.49  723.13
#>   -14p75.24p75.ESH.109 1269.9243 1505.02 1796.76 2095.48 2369.60
#>                       t
#> x.y.country.cell         y2045   y2055   y2065   y2075   y2085   y2095
#>   -14p75.20p25.MRT.100 1659.73 1924.19 2172.30 2387.96 2560.32 2671.07
#>   -14p75.20p75.MRT.101 1671.94 1691.24 1719.25 1765.77 1832.31 1918.47
#>   -14p75.21p25.MRT.102  618.97  619.37  618.74  622.03  630.52  642.20
#>   -14p75.21p75.ESH.103  315.72  317.36  319.61  322.30  327.08  332.39
#>   -14p75.22p25.ESH.104  880.98  939.44  989.54 1035.25 1079.39 1117.61
#>   -14p75.22p75.ESH.105  790.61  871.83  942.81 1002.56 1052.19 1088.94
#>   -14p75.23p25.ESH.106  431.12  448.70  465.22  481.39  494.88  505.93
#>   -14p75.23p75.ESH.107  159.70  160.45  160.95  161.47  163.53  166.31
#>   -14p75.24p25.ESH.108  767.30  798.68  819.21  834.31  844.38  843.52
#>   -14p75.24p75.ESH.109 2600.68 2783.75 2920.70 3006.60 3040.10 3007.86
#>                       t
#> x.y.country.cell         y2105   y2115   y2125   y2135   y2145
#>   -14p75.20p25.MRT.100 2708.86 2708.86 2708.86 2708.86 2708.86
#>   -14p75.20p75.MRT.101 1965.05 1965.05 1965.05 1965.05 1965.05
#>   -14p75.21p25.MRT.102  648.98  648.98  648.98  648.98  648.98
#>   -14p75.21p75.ESH.103  334.66  334.66  334.66  334.66  334.66
#>   -14p75.22p25.ESH.104 1134.64 1134.64 1134.64 1134.64 1134.64
#>   -14p75.22p75.ESH.105 1103.31 1103.31 1103.31 1103.31 1103.31
#>   -14p75.23p25.ESH.106  511.41  511.41  511.41  511.41  511.41
#>   -14p75.23p75.ESH.107  167.49  167.49  167.49  167.49  167.49
#>   -14p75.24p25.ESH.108  839.53  839.53  839.53  839.53  839.53
#>   -14p75.24p75.ESH.109 2972.39 2972.39 2972.39 2972.39 2972.39
#> 
#> , , scenario = B1
#> 
#>                       t
#> x.y.country.cell           y1995   y2005   y2015   y2025   y2035
#>   -14p75.20p25.MRT.100  552.6664  721.85  932.04 1118.33 1267.33
#>   -14p75.20p75.MRT.101 1280.6350 1429.26 1499.74 1531.12 1518.73
#>   -14p75.21p25.MRT.102  554.4384  587.21  603.63  613.98  619.48
#>   -14p75.21p75.ESH.103  276.3431  296.84  305.26  309.78  311.47
#>   -14p75.22p25.ESH.104  451.9981  552.79  623.20  681.60  723.44
#>   -14p75.22p75.ESH.105  277.7437  398.92  502.51  598.73  682.80
#>   -14p75.23p25.ESH.106  292.1132  325.04  349.85  376.11  399.68
#>   -14p75.23p75.ESH.107  133.7772  153.07  157.37  159.07  159.51
#>   -14p75.24p25.ESH.108  383.2277  530.67  590.42  639.68  674.98
#>   -14p75.24p75.ESH.109 1269.9243 1475.64 1687.80 1870.96 1999.15
#>                       t
#> x.y.country.cell         y2045   y2055   y2065   y2075   y2085   y2095
#>   -14p75.20p25.MRT.100 1383.24 1469.16 1510.27 1505.16 1454.54 1361.24
#>   -14p75.20p75.MRT.101 1463.68 1370.97 1257.23 1139.25 1021.52  904.61
#>   -14p75.21p25.MRT.102  617.12  606.77  592.52  579.18  567.73  554.61
#>   -14p75.21p75.ESH.103  309.03  301.99  292.46  281.39  269.77  257.52
#>   -14p75.22p25.ESH.104  747.70  753.98  743.05  718.79  683.68  637.69
#>   -14p75.22p75.ESH.105  754.14  811.59  849.11  865.89  861.01  831.23
#>   -14p75.23p25.ESH.106  418.70  434.27  449.98  468.05  486.99  503.86
#>   -14p75.23p75.ESH.107  158.10  155.21  151.86  148.08  144.47  140.82
#>   -14p75.24p25.ESH.108  692.45  689.79  668.98  634.64  590.05  536.24
#>   -14p75.24p75.ESH.109 2072.68 2090.96 2049.18 1953.77 1811.83 1629.07
#>                       t
#> x.y.country.cell         y2105   y2115   y2125   y2135   y2145
#>   -14p75.20p25.MRT.100 1304.59 1304.59 1304.59 1304.59 1304.59
#>   -14p75.20p75.MRT.101  846.50  846.50  846.50  846.50  846.50
#>   -14p75.21p25.MRT.102  547.06  547.06  547.06  547.06  547.06
#>   -14p75.21p75.ESH.103  251.04  251.04  251.04  251.04  251.04
#>   -14p75.22p25.ESH.104  611.88  611.88  611.88  611.88  611.88
#>   -14p75.22p75.ESH.105  809.60  809.60  809.60  809.60  809.60
#>   -14p75.23p25.ESH.106  511.44  511.44  511.44  511.44  511.44
#>   -14p75.23p75.ESH.107  138.80  138.80  138.80  138.80  138.80
#>   -14p75.24p25.ESH.108  507.06  507.06  507.06  507.06  507.06
#>   -14p75.24p75.ESH.109 1528.15 1528.15 1528.15 1528.15 1528.15
#> 

 getCoords(y) <- map[100:109, c("lon", "lat")]
 getItems(y, "i") <- NULL
 addLocation(y, format = "iso.grid")
#> , , scenario = A2
#> 
#>          t
#> iso.grid      y1995   y2005   y2015   y2025   y2035   y2045   y2055
#>   MRT.100  552.6664  696.44  889.18 1124.11 1389.33 1659.73 1924.19
#>   MRT.101 1280.6350 1429.53 1518.46 1592.09 1640.95 1671.94 1691.24
#>   MRT.102  554.4384  582.36  593.76  605.27  614.58  618.97  619.37
#>   ESH.103  276.3431  295.38  302.62  308.59  313.30  315.72  317.36
#>   ESH.104  451.9981  558.29  646.02  733.13  812.69  880.98  939.44
#>   ESH.105  277.7437  390.18  489.22  596.13  698.33  790.61  871.83
#>   ESH.106  292.1132  326.09  353.25  382.53  409.44  431.12  448.70
#>   ESH.107  133.7772  152.00  155.27  157.35  158.81  159.70  160.45
#>   ESH.108  383.2277  534.73  604.94  668.49  723.13  767.30  798.68
#>   ESH.109 1269.9243 1505.02 1796.76 2095.48 2369.60 2600.68 2783.75
#>          t
#> iso.grid    y2065   y2075   y2085   y2095   y2105   y2115   y2125
#>   MRT.100 2172.30 2387.96 2560.32 2671.07 2708.86 2708.86 2708.86
#>   MRT.101 1719.25 1765.77 1832.31 1918.47 1965.05 1965.05 1965.05
#>   MRT.102  618.74  622.03  630.52  642.20  648.98  648.98  648.98
#>   ESH.103  319.61  322.30  327.08  332.39  334.66  334.66  334.66
#>   ESH.104  989.54 1035.25 1079.39 1117.61 1134.64 1134.64 1134.64
#>   ESH.105  942.81 1002.56 1052.19 1088.94 1103.31 1103.31 1103.31
#>   ESH.106  465.22  481.39  494.88  505.93  511.41  511.41  511.41
#>   ESH.107  160.95  161.47  163.53  166.31  167.49  167.49  167.49
#>   ESH.108  819.21  834.31  844.38  843.52  839.53  839.53  839.53
#>   ESH.109 2920.70 3006.60 3040.10 3007.86 2972.39 2972.39 2972.39
#>          t
#> iso.grid    y2135   y2145
#>   MRT.100 2708.86 2708.86
#>   MRT.101 1965.05 1965.05
#>   MRT.102  648.98  648.98
#>   ESH.103  334.66  334.66
#>   ESH.104 1134.64 1134.64
#>   ESH.105 1103.31 1103.31
#>   ESH.106  511.41  511.41
#>   ESH.107  167.49  167.49
#>   ESH.108  839.53  839.53
#>   ESH.109 2972.39 2972.39
#> 
#> , , scenario = B1
#> 
#>          t
#> iso.grid      y1995   y2005   y2015   y2025   y2035   y2045   y2055
#>   MRT.100  552.6664  721.85  932.04 1118.33 1267.33 1383.24 1469.16
#>   MRT.101 1280.6350 1429.26 1499.74 1531.12 1518.73 1463.68 1370.97
#>   MRT.102  554.4384  587.21  603.63  613.98  619.48  617.12  606.77
#>   ESH.103  276.3431  296.84  305.26  309.78  311.47  309.03  301.99
#>   ESH.104  451.9981  552.79  623.20  681.60  723.44  747.70  753.98
#>   ESH.105  277.7437  398.92  502.51  598.73  682.80  754.14  811.59
#>   ESH.106  292.1132  325.04  349.85  376.11  399.68  418.70  434.27
#>   ESH.107  133.7772  153.07  157.37  159.07  159.51  158.10  155.21
#>   ESH.108  383.2277  530.67  590.42  639.68  674.98  692.45  689.79
#>   ESH.109 1269.9243 1475.64 1687.80 1870.96 1999.15 2072.68 2090.96
#>          t
#> iso.grid    y2065   y2075   y2085   y2095   y2105   y2115   y2125
#>   MRT.100 1510.27 1505.16 1454.54 1361.24 1304.59 1304.59 1304.59
#>   MRT.101 1257.23 1139.25 1021.52  904.61  846.50  846.50  846.50
#>   MRT.102  592.52  579.18  567.73  554.61  547.06  547.06  547.06
#>   ESH.103  292.46  281.39  269.77  257.52  251.04  251.04  251.04
#>   ESH.104  743.05  718.79  683.68  637.69  611.88  611.88  611.88
#>   ESH.105  849.11  865.89  861.01  831.23  809.60  809.60  809.60
#>   ESH.106  449.98  468.05  486.99  503.86  511.44  511.44  511.44
#>   ESH.107  151.86  148.08  144.47  140.82  138.80  138.80  138.80
#>   ESH.108  668.98  634.64  590.05  536.24  507.06  507.06  507.06
#>   ESH.109 2049.18 1953.77 1811.83 1629.07 1528.15 1528.15 1528.15
#>          t
#> iso.grid    y2135   y2145
#>   MRT.100 1304.59 1304.59
#>   MRT.101  846.50  846.50
#>   MRT.102  547.06  547.06
#>   ESH.103  251.04  251.04
#>   ESH.104  611.88  611.88
#>   ESH.105  809.60  809.60
#>   ESH.106  511.44  511.44
#>   ESH.107  138.80  138.80
#>   ESH.108  507.06  507.06
#>   ESH.109 1528.15 1528.15
#> 

 getCells(z) <- map$celliso[100:109]
 addLocation(z)
#> , , scenario = A2
#> 
#>                       t
#> i.region.x.y               y1995   y2005   y2015   y2025   y2035
#>   MRT.100.-14p75.20p25  552.6664  696.44  889.18 1124.11 1389.33
#>   MRT.101.-14p75.20p75 1280.6350 1429.53 1518.46 1592.09 1640.95
#>   MRT.102.-14p75.21p25  554.4384  582.36  593.76  605.27  614.58
#>   ESH.103.-14p75.21p75  276.3431  295.38  302.62  308.59  313.30
#>   ESH.104.-14p75.22p25  451.9981  558.29  646.02  733.13  812.69
#>   ESH.105.-14p75.22p75  277.7437  390.18  489.22  596.13  698.33
#>   ESH.106.-14p75.23p25  292.1132  326.09  353.25  382.53  409.44
#>   ESH.107.-14p75.23p75  133.7772  152.00  155.27  157.35  158.81
#>   ESH.108.-14p75.24p25  383.2277  534.73  604.94  668.49  723.13
#>   ESH.109.-14p75.24p75 1269.9243 1505.02 1796.76 2095.48 2369.60
#>                       t
#> i.region.x.y             y2045   y2055   y2065   y2075   y2085   y2095
#>   MRT.100.-14p75.20p25 1659.73 1924.19 2172.30 2387.96 2560.32 2671.07
#>   MRT.101.-14p75.20p75 1671.94 1691.24 1719.25 1765.77 1832.31 1918.47
#>   MRT.102.-14p75.21p25  618.97  619.37  618.74  622.03  630.52  642.20
#>   ESH.103.-14p75.21p75  315.72  317.36  319.61  322.30  327.08  332.39
#>   ESH.104.-14p75.22p25  880.98  939.44  989.54 1035.25 1079.39 1117.61
#>   ESH.105.-14p75.22p75  790.61  871.83  942.81 1002.56 1052.19 1088.94
#>   ESH.106.-14p75.23p25  431.12  448.70  465.22  481.39  494.88  505.93
#>   ESH.107.-14p75.23p75  159.70  160.45  160.95  161.47  163.53  166.31
#>   ESH.108.-14p75.24p25  767.30  798.68  819.21  834.31  844.38  843.52
#>   ESH.109.-14p75.24p75 2600.68 2783.75 2920.70 3006.60 3040.10 3007.86
#>                       t
#> i.region.x.y             y2105   y2115   y2125   y2135   y2145
#>   MRT.100.-14p75.20p25 2708.86 2708.86 2708.86 2708.86 2708.86
#>   MRT.101.-14p75.20p75 1965.05 1965.05 1965.05 1965.05 1965.05
#>   MRT.102.-14p75.21p25  648.98  648.98  648.98  648.98  648.98
#>   ESH.103.-14p75.21p75  334.66  334.66  334.66  334.66  334.66
#>   ESH.104.-14p75.22p25 1134.64 1134.64 1134.64 1134.64 1134.64
#>   ESH.105.-14p75.22p75 1103.31 1103.31 1103.31 1103.31 1103.31
#>   ESH.106.-14p75.23p25  511.41  511.41  511.41  511.41  511.41
#>   ESH.107.-14p75.23p75  167.49  167.49  167.49  167.49  167.49
#>   ESH.108.-14p75.24p25  839.53  839.53  839.53  839.53  839.53
#>   ESH.109.-14p75.24p75 2972.39 2972.39 2972.39 2972.39 2972.39
#> 
#> , , scenario = B1
#> 
#>                       t
#> i.region.x.y               y1995   y2005   y2015   y2025   y2035
#>   MRT.100.-14p75.20p25  552.6664  721.85  932.04 1118.33 1267.33
#>   MRT.101.-14p75.20p75 1280.6350 1429.26 1499.74 1531.12 1518.73
#>   MRT.102.-14p75.21p25  554.4384  587.21  603.63  613.98  619.48
#>   ESH.103.-14p75.21p75  276.3431  296.84  305.26  309.78  311.47
#>   ESH.104.-14p75.22p25  451.9981  552.79  623.20  681.60  723.44
#>   ESH.105.-14p75.22p75  277.7437  398.92  502.51  598.73  682.80
#>   ESH.106.-14p75.23p25  292.1132  325.04  349.85  376.11  399.68
#>   ESH.107.-14p75.23p75  133.7772  153.07  157.37  159.07  159.51
#>   ESH.108.-14p75.24p25  383.2277  530.67  590.42  639.68  674.98
#>   ESH.109.-14p75.24p75 1269.9243 1475.64 1687.80 1870.96 1999.15
#>                       t
#> i.region.x.y             y2045   y2055   y2065   y2075   y2085   y2095
#>   MRT.100.-14p75.20p25 1383.24 1469.16 1510.27 1505.16 1454.54 1361.24
#>   MRT.101.-14p75.20p75 1463.68 1370.97 1257.23 1139.25 1021.52  904.61
#>   MRT.102.-14p75.21p25  617.12  606.77  592.52  579.18  567.73  554.61
#>   ESH.103.-14p75.21p75  309.03  301.99  292.46  281.39  269.77  257.52
#>   ESH.104.-14p75.22p25  747.70  753.98  743.05  718.79  683.68  637.69
#>   ESH.105.-14p75.22p75  754.14  811.59  849.11  865.89  861.01  831.23
#>   ESH.106.-14p75.23p25  418.70  434.27  449.98  468.05  486.99  503.86
#>   ESH.107.-14p75.23p75  158.10  155.21  151.86  148.08  144.47  140.82
#>   ESH.108.-14p75.24p25  692.45  689.79  668.98  634.64  590.05  536.24
#>   ESH.109.-14p75.24p75 2072.68 2090.96 2049.18 1953.77 1811.83 1629.07
#>                       t
#> i.region.x.y             y2105   y2115   y2125   y2135   y2145
#>   MRT.100.-14p75.20p25 1304.59 1304.59 1304.59 1304.59 1304.59
#>   MRT.101.-14p75.20p75  846.50  846.50  846.50  846.50  846.50
#>   MRT.102.-14p75.21p25  547.06  547.06  547.06  547.06  547.06
#>   ESH.103.-14p75.21p75  251.04  251.04  251.04  251.04  251.04
#>   ESH.104.-14p75.22p25  611.88  611.88  611.88  611.88  611.88
#>   ESH.105.-14p75.22p75  809.60  809.60  809.60  809.60  809.60
#>   ESH.106.-14p75.23p25  511.44  511.44  511.44  511.44  511.44
#>   ESH.107.-14p75.23p75  138.80  138.80  138.80  138.80  138.80
#>   ESH.108.-14p75.24p25  507.06  507.06  507.06  507.06  507.06
#>   ESH.109.-14p75.24p75 1528.15 1528.15 1528.15 1528.15 1528.15
#>